# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Remembra API
  version: "1.0.0"
  summary: External memory for AI assistants over the stable /api/v1 namespace.
  description: |
    Remembra stores memories with explicit type, scope, provenance, and trust
    semantics and exposes them through HTTP, MCP, and SDK clients.

    This document describes the versioned `/api/v1` surface. The legacy routes
    remain available and are not documented here; see `docs/public-api.md`.

    Authentication happens before rate limiting. Send `x-api-key` on a
    protected route. A host that authenticates without a static API key must
    resolve a trusted tenant context itself and, for keyed batches, provide a
    post-authentication credential scope.
  license:
    name: MIT
    url: https://github.com/Hilbras/Remembra/blob/main/LICENSE
  contact:
    name: Hilbras
    url: https://github.com/Hilbras/Remembra/issues
servers:
  - url: http://127.0.0.1:8787
    description: Local default
  - url: https://memory.example.com
    description: TLS deployment (set REMEMBRA_API_KEY)
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/memories:
    post:
      tags: [memories]
      operationId: storeMemory
      summary: Store a memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreRequest"
      responses:
        "200":
          $ref: "#/components/responses/MemoryStored"
        "400":
          $ref: "#/components/responses/InvalidInput"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
    get:
      tags: [memories]
      operationId: listMemories
      summary: List memories with opaque cursor pagination
      parameters:
        - $ref: "#/components/parameters/Scope"
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/MemoryType"
        - name: cursor
          in: query
          description: Opaque keyset cursor. Do not combine with offset.
          schema:
            type: string
            maxLength: 512
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
        - $ref: "#/components/parameters/Limit"
        - name: includeArchived
          in: query
          schema:
            type: boolean
        - name: includeQuarantined
          in: query
          schema:
            type: boolean
        - name: includeExpired
          in: query
          schema:
            type: boolean
        - name: includeFuture
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: One page of memories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Static API key. Rate limiting is charged after authentication.
  parameters:
...
    Scope:
      name: scope
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
...
  responses:
    MemoryStored:
      description: Stored memory
      content:
        application/json:
          schema:
            type: object
            required: [id, message]
            properties:
              id:
                type: string
              message:
                type: string
              memory:
                $ref: "#/components/schemas/Memory"
    InvalidInput:
      description: Validation failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimited:
      description: Rate limit exceeded after authentication
      headers:
        retry-after:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    MemoryType:
      type: string
      enum:
        [
          "fact",
          "preference",
          "decision",
          "constraint",
          "instruction",
          "role",
          "entity",
          "relationship",
          "observation",
          "episode",
          "pattern",
        ]
    TrustLevel:
      type: string
      enum: [unverified, trusted, verified, system]
    RetentionMode:
      type: string
      enum: [decaying, pinned, neverExpire, ephemeral]
    Memory:
      type: object
      required: [id, type, content, scope, version, createdAt, updatedAt]
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/MemoryType"
        content:
          type: string
        scope:
          type: string
        tags:
          type: array
          items:
            type: string
        importance:
          type: integer
          minimum: 1
          maximum: 10
        confidence:
          type: number
        trust:
          $ref: "#/components/schemas/TrustLevel"
        owner:
          type: string
        access:
          type: string
        retention:
          $ref: "#/components/schemas/RetentionMode"
        version:
          type: integer
          minimum: 1
        provenance:
          type: object
          additionalProperties: true
        relations:
          type: array
          items:
            type: object
        createdAt:
          type: string
        updatedAt:
          type: string
        embedding:
          description: Never present in a public response.
          type: array
          items:
            type: number
    StoreRequest:
      type: object
      required: [type, content]
      additionalProperties: false
      description: |
        Server-managed fields (`id`, `tenantId`, `organizationId`, `userId`,
        `agentId`, `embedding`) are rejected.
      properties:
        type:
          $ref: "#/components/schemas/MemoryType"
        content:
          type: string
          minLength: 1
        scope:
          type: string
        tags:
          type: array
          items:
            type: string
        importance:
          type: integer
          minimum: 1
          maximum: 10
        confidence:
          type: number
        trust:
          $ref: "#/components/schemas/TrustLevel"
        owner:
          type: string
        access:
          type: string
        retention:
          $ref: "#/components/schemas/RetentionMode"
        source:
          type: string
        validFrom:
          type: string
        validUntil:
          type: string
        meta:
          type: object
          additionalProperties: true
        provenance:
          type: object
          additionalProperties: true
...
    MemoryPage:
      type: object
      properties:
        memories:
          type: array
          items:
            $ref: "#/components/schemas/Memory"
        count:
          type: integer
        next_cursor:
          type: string
          nullable: true
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
          description: Machine-readable code; legacy responses may omit it.